% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_df.R
\name{ci_df}
\alias{ci_df}
\title{Calculate mean with confidence intervals.}
\usage{
ci_df(.data, .summary_var, ..., ci = 0.95, groups_col = FALSE)
}
\arguments{
\item{.data}{A data frame}

\item{.summary_var}{Variable that is being summarized}

\item{...}{Grouping variable(s)}

\item{ci}{Confidence interval. Defaults to 0.95}

\item{groups_col}{List grouping variable under column named 'group'. Defaults
to FALSE.}
}
\value{
A data frame with variables `mean`, `ci_low`, `ci_high`, `sd`, `n`.
If the data frame is grouped, grouping variable will also be listed.
}
\description{
Calculate mean with confidence intervals.
}
\examples{
ci_df(mtcars, mpg, cyl, gear)
}
\seealso{
[Hmisc::smean.cl.normal()] which this function utilizes
}
