% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_default_dirs.R
\name{create_default_dirs}
\alias{create_default_dirs}
\title{Create default directories I use for projects}
\usage{
create_default_dirs()
}
\description{
Create default directories I use for projects
}
