% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_kable.R
\name{custom_kable}
\alias{custom_kable}
\title{custom_kable: A customized format using for printing tables}
\usage{
custom_kable(
  data,
  digits = 3,
  caption = NA,
  col.names = NA,
  row.names = NA,
  format = NULL,
  escape = TRUE,
  format.args = list()
)
}
\arguments{
\item{data}{A data frame or tibble. Matrices usually work, except when they don't.}

\item{digits}{Maximum number of digits for numeric columns, passed to
\code{round()}. This can also be a vector of length \code{ncol(x)}, to set
the number of digits for individual columns.}

\item{caption}{The table caption.}

\item{col.names}{A character vector of column names to be used in the table.}

\item{row.names}{Logical: whether to include row names. By default, row names
are included if \code{rownames(x)} is neither \code{NULL} nor identical to
\code{1:nrow(x)}.}

\item{format}{A character string. Possible values are \code{latex},
\code{html}, \code{markdown}, \code{pandoc}, and \code{rst}; this will be
automatically determined if the function is called within \pkg{knitr}; it
can also be set in the global option \code{knitr.table.format}. If
\code{format} is a function, it must return a character string.}

\item{escape}{Boolean; whether to escape special characters when producing
HTML or LaTeX tables. When \code{escape = FALSE}, you have to make sure
that special characters will not trigger syntax errors in LaTeX or HTML.}

\item{format.args}{A list of arguments to be passed to \code{\link{format}()}
to format table values, e.g. \code{list(big.mark = ',')}.}
}
\value{
A character vector of the table source code.
}
\description{
This function uses \code{knitr::kable()} and \code{kableExtra::kable_styling} to build
a customized table that, in my opinion, looks better than the default.
}
\details{
Missing values (\code{NA}) in the table are displayed as \code{NA} by
default. If you want to display them with other characters, you can set the
option \code{knitr.kable.NA}, e.g. \code{options(knitr.kable.NA = '')} to
hide \code{NA} values.
}
\examples{
custom_kable(df, caption = "A caption")
}
\references{
See
  \url{https://github.com/yihui/knitr-examples/blob/master/091-knitr-table.Rnw}
   for some examples in LaTeX, but they also apply to other document formats.
}
\seealso{
See \code{\link[knitr]{kable}} and \code{\link[kableExtra]{kable_styling}}
}
