% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_pval.R
\name{format_pval}
\alias{format_pval}
\title{Format p-value}
\usage{
format_pval(p, digits = 3)
}
\arguments{
\item{p}{One or more p-values as either a numeric vector or numeric variable in a data frame.}

\item{digits}{How many digits to round to? Defaults to 3.}
}
\value{
A character vector of rounded p-values. Any p-values less than .001 are shown as "< .001"
}
\description{
Format p-value
}
\examples{
p.value <- c(.01, .005, .00024)
format_pval(p.value)
}
