% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh.R
\name{hh}
\alias{hh}
\title{Get first 5 rows and first 5 columns of a data frame or matrix}
\usage{
hh(d)
}
\arguments{
\item{d}{A data frame or matrix}
}
\description{
Get first 5 rows and first 5 columns of a data frame or matrix
}
\examples{
hh(d)
}
