% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_load_packages.R
\name{install_load_packages}
\alias{install_load_packages}
\title{Install and load packages

This function will check for each package in the user-ddefined and install
them if they are missing, and then load them. It also takes a list of github
packages and their repos.}
\usage{
install_load_packages()
}
\arguments{
\item{packages}{A list of the packages you want to load}

\item{github_packages}{A list of the github packages and their repos.
\code{package} = package names, \code{repo} = package repos}
}
\description{
Install and load packages

This function will check for each package in the user-ddefined and install
them if they are missing, and then load them. It also takes a list of github
packages and their repos.
}
\examples{
packages <- c("tidyverse", "here", "lubridate")
github_packages <- list(package = c("fancyCorr", "report"), repo = c("sbw78/fancyCorr", "easystats/report"))
The function pulls the variables from the global environment after they are defined
install_load_packages()
}
\author{
Bryan West
}
\keyword{github}
\keyword{packages,}
