% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_forest_plot.R
\name{make_forest_plot}
\alias{make_forest_plot}
\title{Create a forest plot}
\usage{
make_forest_plot(
  d,
  .measure_var,
  .group_var,
  xlab = NULL,
  ylab = NULL,
  ci_low = ci_low,
  ci_high = ci_high,
  xintercept = NULL
)
}
\arguments{
\item{d}{A data frame in long format that has boundary values for a confidence interval}

\item{.measure_var}{Numeric variable to be plotted on the x-axis}

\item{.group_var}{Factor or character variable to differentiate plots}

\item{xlab}{Label for x-axis}

\item{ylab}{Label for y-axis}

\item{ci_low}{Minimum confidence interval variable. Defaults to `ci_low`}

\item{ci_high}{Maximum confidence interval variable Defaults to `ci_high`}

\item{xintercept}{Value of x-intercept for vertical dotted line}
}
\value{
A ggplot plot
}
\description{
Create a forest plot
}
\examples{
forest_plot
}
\seealso{
\code{\link[tidyr]{pivot_longer}}, \code{\link[ggplot2]{ggplot}}
}
