% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_build.R
\name{regex_build}
\alias{regex_build}
\title{Create a regular expression from a character vector or list}
\usage{
regex_build(list, modifier = "single")
}
\arguments{
\item{list}{A list of characters}

\item{modifier}{Type of regex. 'single' will break every word into a separate
section of the expression. 'multi' will allow phrases to be components of the
expression. Defaults to 'single'.}
}
\value{
A regular expression
}
\description{
Create a regular expression from a character vector or list
}
\examples{
regex_build(c("apple", "banana", "carrot"))
regex_build(c("dog treats", "cereal", "Boston baked beans"))
}
