% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{A function to get the standard error}
\usage{
se(x, na.rm = FALSE)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Remove NAs? Defaults to FALSE.}
}
\description{
A function to get the standard error
}
\examples{
se(x, na.rm = TRUE)
}
