% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_packages.R
\name{use_packages}
\alias{use_packages}
\title{Create R/packages.R file}
\usage{
use_packages(open = interactive())
}
\arguments{
\item{open}{Logical value, whether to open packages.R for editing}
}
\value{

}
\description{
Create R/packages.R file
}
\author{
Bryan West
}
